+++ waspc/data/Generator/templates/sdk/wasp/client/app/router/router.tsx
@@ -14,7 +14,7 @@ export function getRouter({
   routesMapping,
   rootElement,
 }: {
-  routesMapping: Record<string, () => Promise<{ Component: React.ComponentType }>>,
+  routesMapping: Record<string, { lazy: () => Promise<{ Component: React.ComponentType }> } | { Component: React.ComponentType }>,
   rootElement: React.ReactNode,
 }) {
   const waspDefinedRoutes = [
@@ -28,7 +28,7 @@ export function getRouter({
   const userDefinedRoutes = Object.entries(routes).map(([routeKey, route]) => {
     return {
       path: route.to,
-      lazy: routesMapping[routeKey],
+      ...routesMapping[routeKey],
     }
   })