+++ waspc/data/Generator/templates/sdk/wasp/client/vite/virtual-files/files/routes.tsx
@@ -3,15 +3,19 @@
 {=# isAuthEnabled =}
 import { createAuthRequiredPage } from "wasp/client/app"
 {=/ isAuthEnabled =}
+{=# eagerImports =}
+{=& importStatement =}
+{=/ eagerImports =}
 
 export const routesMapping = {
   {=# routes =}
-  {= name =}: async () => {
+  {=# isLazy =}
+  {= name =}: { lazy: async () => {
     {=# isDefaultExport =}
     const { default: Component } = await import('{=& importPath =}')
     {=/ isDefaultExport =}
     {=^ isDefaultExport =}
-    const { {= exportedName =}: Component } = await import('{=& importPath =}')
+    const { {= importIdentifier =}: Component } = await import('{=& importPath =}')
     {=/ isDefaultExport =}
     {=# isAuthRequired =}
     return { Component: createAuthRequiredPage(Component) }
@@ -19,6 +23,17 @@ export const routesMapping = {
     {=^ isAuthRequired =}
     return { Component }
     {=/ isAuthRequired =}
+  }},
+  {=/ isLazy =}
+  {=^ isLazy =}
+  {= name =}: {
+    {=# isAuthRequired =}
+    Component: createAuthRequiredPage({= importIdentifier =}),
+    {=/ isAuthRequired =}
+    {=^ isAuthRequired =}
+    Component: {= importIdentifier =},
+    {=/ isAuthRequired =}
   },
+  {=/ isLazy =}
   {=/ routes =}
 } as const;