+++ waspc/e2e-tests/test-outputs/snapshots/kitchen-sink-golden/wasp-app/.wasp/out/sdk/wasp/client/vite/virtual-files/files/routes.tsx
@@ -2,80 +2,80 @@
 import { createAuthRequiredPage } from "wasp/client/app"
 
 export const routesMapping = {
-  HomeRoute: async () => {
+  HomeRoute: { lazy: async () => {
     const { HomePage: Component } = await import('./src/pages/HomePage')
     return { Component }
-  },
-  CatchAllRoute: async () => {
+  }},
+  CatchAllRoute: { lazy: async () => {
     const { CatchAllPage: Component } = await import('./src/pages/CatchAllPage')
     return { Component }
-  },
-  SignupRoute: async () => {
+  }},
+  SignupRoute: { lazy: async () => {
     const { default: Component } = await import('./src/features/auth/pages/Signup')
     return { Component }
-  },
-  LoginRoute: async () => {
+  }},
+  LoginRoute: { lazy: async () => {
     const { default: Component } = await import('./src/features/auth/pages/Login')
     return { Component }
-  },
-  PasswordResetRoute: async () => {
+  }},
+  PasswordResetRoute: { lazy: async () => {
     const { PasswordReset: Component } = await import('./src/features/auth/pages/PasswordReset')
     return { Component }
-  },
-  EmailVerificationRoute: async () => {
+  }},
+  EmailVerificationRoute: { lazy: async () => {
     const { EmailVerification: Component } = await import('./src/features/auth/pages/EmailVerification')
     return { Component }
-  },
-  RequestPasswordResetRoute: async () => {
+  }},
+  RequestPasswordResetRoute: { lazy: async () => {
     const { RequestPasswordReset: Component } = await import('./src/features/auth/pages/RequestPasswordReset')
     return { Component }
-  },
-  ProfileRoute: async () => {
+  }},
+  ProfileRoute: { lazy: async () => {
     const { ProfilePage: Component } = await import('./src/features/auth/pages/ProfilePage')
     return { Component: createAuthRequiredPage(Component) }
-  },
-  ManualSignupRoute: async () => {
+  }},
+  ManualSignupRoute: { lazy: async () => {
     const { ManualSignupPage: Component } = await import('./src/features/auth/pages/ManualSignupPage')
     return { Component }
-  },
-  CustomSignupRoute: async () => {
+  }},
+  CustomSignupRoute: { lazy: async () => {
     const { CustomSignupPage: Component } = await import('./src/features/auth/pages/CustomSignupPage')
     return { Component }
-  },
-  TasksRoute: async () => {
+  }},
+  TasksRoute: { lazy: async () => {
     const { TasksPage: Component } = await import('./src/features/operations/pages/TasksPage')
     return { Component: createAuthRequiredPage(Component) }
-  },
-  TaskRoute: async () => {
+  }},
+  TaskRoute: { lazy: async () => {
     const { TaskDetailPage: Component } = await import('./src/features/operations/pages/TaskDetailPage')
     return { Component: createAuthRequiredPage(Component) }
-  },
-  SerializationRoute: async () => {
+  }},
+  SerializationRoute: { lazy: async () => {
     const { SerializationPage: Component } = await import('./src/features/operations/pages/SerializationPage')
     return { Component }
-  },
-  JobsRoute: async () => {
+  }},
+  JobsRoute: { lazy: async () => {
     const { JobsPage: Component } = await import('./src/features/jobs/pages/JobsPage')
     return { Component: createAuthRequiredPage(Component) }
-  },
-  ApisRoute: async () => {
+  }},
+  ApisRoute: { lazy: async () => {
     const { ApisPage: Component } = await import('./src/features/apis/pages/ApisPage')
     return { Component }
-  },
-  CrudListRoute: async () => {
+  }},
+  CrudListRoute: { lazy: async () => {
     const { ListPage: Component } = await import('./src/features/crud/pages/ListPage')
     return { Component: createAuthRequiredPage(Component) }
-  },
-  CrudDetailRoute: async () => {
+  }},
+  CrudDetailRoute: { lazy: async () => {
     const { DetailPage: Component } = await import('./src/features/crud/pages/DetailPage')
     return { Component: createAuthRequiredPage(Component) }
-  },
-  StreamingRoute: async () => {
+  }},
+  StreamingRoute: { lazy: async () => {
     const { StreamingTestPage: Component } = await import('./src/features/streaming/pages/StreamingTestPage')
     return { Component }
-  },
-  ChatRoute: async () => {
+  }},
+  ChatRoute: { lazy: async () => {
     const { ChatPage: Component } = await import('./src/features/chat/pages/ChatPage')
     return { Component: createAuthRequiredPage(Component) }
-  },
+  }},
 } as const;